+++ R/f_interface_lrtest.R
@@ -7,11 +7,12 @@
 #' If not given, the default model names are used.
 #'
 #' @description
-#' \code{lrtest} carrys out likelihood ratio tests to compare nested CLV models.
+#' \code{lrtest} carries out likelihood ratio tests to compare nested CLV models
+#' of the same family that were fitted on the same transaction data.
 #'
-#' The method consecutively compares the first model given in \code{object} with all the
-#' other models passed in \code{...}. An asymptotic likelihood ratio test is carried out:
-#' Twice the difference in log-likelihoods is compared with a Chi-squared distribution.
+#' The method compares each two consecutive models.
+#' An asymptotic likelihood ratio test is carried out: Twice the difference in
+#' log-likelihoods is compared with a Chi-squared distribution.
 #'
 #' @param object An fitted model object inheriting from \code{clv.fitted}.
 #' @param ... Other models objects fitted on the same transaction data