+++ tests/testthat/helper_arrange.R
@@ -19,7 +19,7 @@ fct.helper.load.apparelDynCov <- function(){.load.data.locally("apparelDynCov")}
 
 
 
-fct.helper.create.clvdata.cdnow <- function(data.cdnow = NULL, estimation.split = 37) {
+fct.helper.create.clvdata.cdnow <- function(data.cdnow = NULL, estimation.split = 37, name.price = "Price") {
   if (is.null(data.cdnow)) {
     data.cdnow <- fct.helper.load.cdnow()
   }
@@ -27,14 +27,15 @@ fct.helper.create.clvdata.cdnow <- function(data.cdnow = NULL, estimation.split 
     data.transactions = data.cdnow,
     date.format = "ymd",
     time.unit = "w",
-    estimation.split = estimation.split
+    estimation.split = estimation.split,
+    name.price = name.price
   )
   return(clv.cdnow)
 }
 
 fct.helper.create.clvdata.apparel.nocov <- function(
     data.apparelTrans = NULL,
-    estimation.split = 40) {
+    estimation.split = 104) {
 
   if (is.null(data.apparelTrans)) {
     data.apparelTrans <- fct.helper.load.apparelTrans()
@@ -45,13 +46,13 @@ fct.helper.create.clvdata.apparel.nocov <- function(
     date.format = "ymd",
     time.unit = "W",
     estimation.split = estimation.split
-    ))
+  ))
 }
 
 fct.helper.create.clvdata.apparel.staticcov <- function(
     data.apparelTrans = NULL,
     data.apparelStaticCov = NULL,
-    estimation.split = 40,
+    estimation.split = 104,
     names.cov.life = c("Gender", "Channel"),
     names.cov.trans = c("Gender", "Channel")) {
   if (is.null(data.apparelTrans)) {
@@ -76,9 +77,9 @@ fct.helper.create.clvdata.apparel.staticcov <- function(
 fct.helper.create.clvdata.apparel.dyncov <- function(
     data.apparelTrans = NULL,
     data.apparelDynCov = NULL,
-    estimation.split = 40,
-    names.cov.life = c("Marketing", "Gender", "Channel"),
-    names.cov.trans = c("Marketing", "Gender", "Channel")) {
+    estimation.split = 104,
+    names.cov.life = c("High.Season", "Gender", "Channel"),
+    names.cov.trans = c("High.Season", "Gender", "Channel")) {
 
   if (is.null(data.apparelTrans)) {
     data.apparelTrans <- fct.helper.load.apparelTrans()
@@ -110,12 +111,16 @@ fct.helper.create.clvdata.apparel.dyncov <- function(
 fit.cdnow <- function(
     data.cdnow = NULL,
     estimation.split = 37,
+    name.price = 'Price',
     model = pnbd,
     start.params.model = c(),
     verbose = FALSE,
     optimx.args = list()) {
+
   clv.cdnow <- fct.helper.create.clvdata.cdnow(
-    data.cdnow = data.cdnow, estimation.split = estimation.split
+    data.cdnow = data.cdnow,
+    estimation.split = estimation.split,
+    name.price=name.price
   )
 
   return(do.call(
@@ -133,14 +138,14 @@ fit.cdnow <- function(
 
 fit.apparel.nocov <- function(
     data.apparelTrans = NULL,
-    estimation.split = 40,
+    estimation.split = 104,
     model = pnbd,
     verbose=FALSE,
     # start.params.model = c(),
     # verbose = FALSE,
     # optimx.args = list()
     ...
-    ) {
+) {
 
   clv.data.apparel <- fct.helper.create.clvdata.apparel.nocov(
     data.apparelTrans = data.apparelTrans,
@@ -160,7 +165,7 @@ fit.apparel.nocov <- function(
 fit.apparel.static <- function(
     data.apparelTrans = NULL,
     data.apparelStaticCov = NULL,
-    estimation.split = 40,
+    estimation.split = 104,
     names.cov.life = c("Gender", "Channel"),
     names.cov.trans = c("Gender", "Channel"),
     model = pnbd,
@@ -169,7 +174,7 @@ fit.apparel.static <- function(
     # verbose = FALSE,
     # optimx.args = list(),
     ...
-    ) {
+) {
   clv.data.apparel.cov <- fct.helper.create.clvdata.apparel.staticcov(
     data.apparelTrans = data.apparelTrans,
     data.apparelStaticCov = data.apparelStaticCov,
@@ -192,38 +197,9 @@ fit.apparel.static <- function(
 fit.apparel.dyncov <- function(
     data.apparelTrans = NULL,
     data.apparelDynCov = NULL,
-    estimation.split = 40,
-    names.cov.life = c("Marketing", "Gender", "Channel"),
-    names.cov.trans = c("Marketing", "Gender", "Channel"),
-    model = pnbd,
-    verbose=FALSE,
-    ...
-) {
-  clv.data.apparel.dyncov <- fct.helper.create.clvdata.apparel.dyncov(
-    data.apparelTrans = data.apparelTrans,
-    data.apparelDynCov = data.apparelDynCov,
-    estimation.split = estimation.split,
-    names.cov.life = names.cov.life,
-    names.cov.trans = names.cov.trans
-  )
-
-  return(do.call(
-    what = model,
-    args = list(
-      clv.data = clv.data.apparel.dyncov,
-      verbose=verbose,
-      ...
-    )
-  ))
-}
-
-
-fit.apparel.dyncov <- function(
-    data.apparelTrans = NULL,
-    data.apparelDynCov = NULL,
-    estimation.split = 40,
-    names.cov.life = c("Marketing", "Gender", "Channel"),
-    names.cov.trans = c("Marketing", "Gender", "Channel"),
+    estimation.split = 104,
+    names.cov.life = c("High.Season", "Gender", "Channel"),
+    names.cov.trans = c("High.Season", "Gender", "Channel"),
     model = pnbd,
     verbose=FALSE,
     ...
@@ -268,7 +244,7 @@ fct.helper.dyncov.quickfit <- function(clv.data.dyn, hessian){
     clv.data=clv.data.dyn,
     # start params from std model fitted before apparel dyncov
     # other start params may yield estimated params which are unsuitable for prediction/plot (NAs & Inf) which removes dates during plotting
-    start.params.model = c(r= 0.7579, alpha= 4.7419, s= 0.5432, beta=22.1892),
+    #start.params.model = c(r= 0.7422, alpha= 2.1222, s= 0.9991, beta=92.0867),
     optimx.args = l.quickfit.args,
     verbose = FALSE)
 
@@ -281,9 +257,9 @@ fct.helper.dyncov.quickfit <- function(clv.data.dyn, hessian){
 }
 
 fct.helper.dyncov.quickfit.apparel.data <- function(data.apparelTrans=NULL, data.apparelDynCov=NULL, hessian=FALSE,
-                                                    estimation.split=40,
-                                                    names.cov.life = c("Marketing", "Gender", "Channel"),
-                                                    names.cov.trans = c("Marketing", "Gender", "Channel")){
+                                                    estimation.split=104,
+                                                    names.cov.life = c("High.Season", "Gender", "Channel"),
+                                                    names.cov.trans = c("High.Season", "Gender", "Channel")){
   clv.apparel.dyn <- fct.helper.create.clvdata.apparel.dyncov(
     data.apparelTrans=data.apparelTrans,
     data.apparelDynCov=data.apparelDynCov,
@@ -306,8 +282,8 @@ fct.helper.dyncov.quickfit.apparel.data <- function(data.apparelTrans=NULL, data
   #                                             123.954642184586, -79.6231571791094, -8.32942983004706,  6.37531270788257, -0.205974188766537, -6.37531003596571, -1.27313616513791, 29.2516552513793, 79.6231520568093, 42.6967589924783,
   #                                             81.2903073168579, -49.7391927152517, -3.62426045915698, 2.4410199256767, 0.0130428435022102, -1.27313681231238, -2.44101785877623, 17.3123174944356, 42.6967589924783, 49.7391916698474),
   #                                           .Dim = c(10L, 10L),
-  #                                           .Dimnames = list(c("log.r", "log.alpha", "log.s", "log.beta", "life.Marketing", "life.Gender", "life.Channel", "trans.Marketing", "trans.Gender", "trans.Channel"),
-  #                                                            c("log.r",  "log.alpha", "log.s", "log.beta", "life.Marketing", "life.Gender", "life.Channel", "trans.Marketing", "trans.Gender", "trans.Channel")))
+  #                                           .Dimnames = list(c("log.r", "log.alpha", "log.s", "log.beta", "life.High.Season", "life.Gender", "life.Channel", "trans.High.Season", "trans.Gender", "trans.Channel"),
+  #                                                            c("log.r",  "log.alpha", "log.s", "log.beta", "life.High.Season", "life.Gender", "life.Channel", "trans.High.Season", "trans.Gender", "trans.Channel")))
 
   return(fitted.dyncov)
 }
@@ -323,7 +299,7 @@ fct.helper.default.newcustomer.covdata.static <- function(){
 
   return(data.frame(
     Gender=1,
-    Channel=6.78
+    Channel=2
   ))
 }
 
@@ -337,5 +313,5 @@ fct.helper.default.newcustomer.covdata.dyncov <- function(){
     Cov.Date=cov.dates,
     Gender=rep_len(0, length(cov.dates)),
     Channel=rep_len(c(-0.678, 0, 2, 1.23, -1.23, -2), length(cov.dates)),
-    Marketing=rep_len(c(4, 0, 7, 2, 9, 0), length(cov.dates))))
+    High.Season=rep_len(c(4, 0, 7, 2, 9, 0), length(cov.dates))))
 }